requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
appdirs==1.4.3
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.2.0
ccxt==1.54.66
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click-datetime==0.2
colorama==0.4.4
cryptography==3.4.7
cycler==0.10.0
Django==2.2.24
django-pandas==0.6.4
ecdsa==0.13
Events==0.3
executing==0.8.0
gitdb==4.0.7
GitPython==3.1.18
icecream==2.1.1
idna==3.2
kiwisolver==1.3.1
matplotlib==3.4.3
multidict==5.1.0
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
plotly==5.2.1
prettytable==0.7.2
pycares==4.0.0
pycparser==2.20
pycryptodome==3.8.0
Pygments==2.9.0
pylibscrypt==1.8.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scrypt==0.8.13
six==1.16.0
smmap==4.0.0
sqlparse==0.4.1
ta==0.7.0
tenacity==8.0.1
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
